+++ crates/domain/src/routine.rs
@@ -811,6 +811,7 @@ mod tests {
         vec![Exercise {
             id: 1.into(),
             name: Name::new("A").unwrap(),
+            notes: String::new(),
             muscles: vec![
                 ExerciseMuscle {
                     muscle_id: MuscleID::Pecs,